+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.17] - 2026-03-26
+
+### Test
+- Update test_planfile_final.py
+
+### Other
+- Update examples/strategy_free_test.yaml
+- Update planfile/models_v2.py
+
 ## [0.1.16] - 2026-03-26
 
 ### Docs